@@ -26,10 +26,10 @@
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
     "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^19.0.4",
     "graphql": "^15.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^7.4.1",
     "react-scripts": "5.0.0",
     "typescript": "^4.4.2",
